+++ ltnc/ast/expr/Char.hxx
@@ -3,11 +3,11 @@
 
 namespace ltn::c::ast::expr {
 	struct Char final : public Expression {
-		Char(std::uint8_t value, const SourceLocation & location);
+		Char(std::uint32_t value, const SourceLocation & location);
 		virtual void accept(const ExprVisitor & visitor) const override;
 
-		std::uint8_t value;
+		std::uint32_t value;
 	};
 
-	std::unique_ptr<Char> character(const SourceLocation & location, std::uint8_t value);
+	std::unique_ptr<Char> character(const SourceLocation & location, std::uint32_t value);
 }
